Added vgg16 and vgg19
