+++ models/resnet.py
@@ -1,0 +1,232 @@
+import torch.nn as nn
+import math
+import numpy as np
+import torch.utils.model_zoo as model_zoo
+import torch
+from torch.autograd import Variable
+
+model_urls = {
+    "resnet18": "https://download.pytorch.org/models/resnet18-5c106cde.pth",
+    "resnet34": "https://download.pytorch.org/models/resnet34-333f7ec4.pth",
+    "resnet50": "https://download.pytorch.org/models/resnet50-19c8e357.pth",
+    "resnet101": "https://download.pytorch.org/models/resnet101-5d3b4d8f.pth",
+    "resnet152": "https://download.pytorch.org/models/resnet152-b121ed2d.pth",
+}
+
+
+def conv3x3(in_planes, out_planes, stride=1):
+    "3x3 convolution with padding"
+    return nn.Conv2d(in_planes, out_planes, kernel_size=3, stride=stride, padding=1, bias=False)
+
+
+class BasicBlock(nn.Module):
+    expansion = 1
+
+    def __init__(self, inplanes, planes, stride=1, downsample=None):
+        super(BasicBlock, self).__init__()
+        self.conv1 = conv3x3(inplanes, planes, stride)
+        self.bn1 = nn.BatchNorm2d(planes)
+        self.relu = nn.ReLU(inplace=True)
+        self.conv2 = conv3x3(planes, planes)
+        self.bn2 = nn.BatchNorm2d(planes)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x):
+        residual = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+
+        if self.downsample is not None:
+            residual = self.downsample(x)
+
+        out += residual
+        out = self.relu(out)
+
+        return out
+
+
+class Bottleneck(nn.Module):
+    expansion = 4
+
+    def __init__(self, inplanes, planes, stride=1, downsample=None):
+        super(Bottleneck, self).__init__()
+        self.conv1 = nn.Conv2d(inplanes, planes, kernel_size=1, bias=False)
+        self.bn1 = nn.BatchNorm2d(planes)
+        self.conv2 = nn.Conv2d(planes, planes, kernel_size=3, stride=stride, padding=1, bias=False)
+        self.bn2 = nn.BatchNorm2d(planes)
+        self.conv3 = nn.Conv2d(planes, planes * 4, kernel_size=1, bias=False)
+        self.bn3 = nn.BatchNorm2d(planes * 4)
+        self.relu = nn.ReLU(inplace=True)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x):
+        residual = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+        out = self.relu(out)
+
+        out = self.conv3(out)
+        out = self.bn3(out)
+
+        if self.downsample is not None:
+            residual = self.downsample(x)
+
+        out += residual
+        out = self.relu(out)
+
+        return out
+
+
+class ResNet(nn.Module):
+    def __init__(self, block, layers, num_classes, channels):
+        self.inplanes = 64
+        super(ResNet, self).__init__()
+        self.conv1_custom = nn.Conv2d(channels, 64, kernel_size=7, stride=2, padding=3, bias=False)
+        self.bn1 = nn.BatchNorm2d(64)
+        self.relu = nn.ReLU(inplace=True)
+        self.maxpool = nn.MaxPool2d(kernel_size=3, stride=2, padding=1)
+        self.layer1 = self._make_layer(block, 64, layers[0])
+        self.layer2 = self._make_layer(block, 128, layers[1], stride=2)
+        self.layer3 = self._make_layer(block, 256, layers[2], stride=2)
+        self.layer4 = self._make_layer(block, 512, layers[3], stride=2)
+        self.avgpool = nn.AvgPool2d(7)
+        self.fc_custom = nn.Linear(512 * block.expansion, num_classes)
+        for m in self.modules():
+            if isinstance(m, nn.Conv2d):
+                n = m.kernel_size[0] * m.kernel_size[1] * m.out_channels
+                m.weight.data.normal_(0, math.sqrt(2.0 / n))
+            elif isinstance(m, nn.BatchNorm2d):
+                m.weight.data.fill_(1)
+                m.bias.data.zero_()
+
+    def _make_layer(self, block, planes, blocks, stride=1):
+        downsample = None
+        if stride != 1 or self.inplanes != planes * block.expansion:
+            downsample = nn.Sequential(
+                nn.Conv2d(
+                    self.inplanes,
+                    planes * block.expansion,
+                    kernel_size=1,
+                    stride=stride,
+                    bias=False,
+                ),
+                nn.BatchNorm2d(planes * block.expansion),
+            )
+
+        layers = []
+        layers.append(block(self.inplanes, planes, stride, downsample))
+        self.inplanes = planes * block.expansion
+        for i in range(1, blocks):
+            layers.append(block(self.inplanes, planes))
+
+        return nn.Sequential(*layers)
+
+    def forward(self, x):
+        x = self.conv1_custom(x)
+        x = self.bn1(x)
+        x = self.relu(x)
+        x = self.maxpool(x)
+
+        x = self.layer1(x)
+        x = self.layer2(x)
+        x = self.layer3(x)
+        x = self.layer4(x)
+
+        x = self.avgpool(x)
+        x = x.view(x.size(0), -1)
+        out = self.fc_custom(x)
+        return out
+
+
+# Transform the original 3 channel weight to the number channels requested
+def cross_modality_pretrain(conv1_weight, channels):
+    # Accumulate conv1 weights across the 3 channels
+    S = 0
+    for i in range(3):
+        S += conv1_weight[:, i, :, :]
+
+    # Get average of weights
+    avg = S / 3.0
+    new_conv1_weight = torch.FloatTensor(64, channels, 7, 7)
+
+    # Assign average weight to each of the channels in conv1
+    for i in range(channels):
+        new_conv1_weight[:, i, :, :] = avg.data
+
+    return new_conv1_weight
+
+
+# Adapt pretrained weights of conv1 to created model according to number of channels
+def weight_transform(model_dict, pretrain_dict, channels):
+    weight_dict = {k: v for k, v in pretrain_dict.items() if k in model_dict}
+    w3 = pretrain_dict["conv1.weight"]
+
+    if channels == 3:
+        wt = w3
+    else:
+        wt = cross_modality_pretrain(w3, channels)
+
+    # Assign pretrained weights to model
+    weight_dict["conv1_custom.weight"] = wt
+    model_dict.update(weight_dict)
+    return model_dict
+
+
+def resnet18(pretrained, num_classes, channels):
+    """Constructs a ResNet-18 model.
+    Args:
+        pretrained (bool): If True, returns a model pre-trained on ImageNet
+    """
+    model = ResNet(BasicBlock, [2, 2, 2, 2], num_classes=num_classes, channels=channels)
+    if pretrained:
+        pretrain_dict = model_zoo.load_url(model_urls["resnet18"])
+        model_dict = model.state_dict()
+        model_dict = weight_transform(model_dict, pretrain_dict, channels)
+        model.load_state_dict(model_dict)
+    return model
+
+
+def resnet50(pretrained, num_classes, channels):
+
+    model = ResNet(Bottleneck, [3, 4, 6, 3], num_classes=num_classes, channels=channels)
+    if pretrained:
+        pretrain_dict = model_zoo.load_url(model_urls["resnet50"])
+        model_dict = model.state_dict()
+        model_dict = weight_transform(model_dict, pretrain_dict, channels)
+        model.load_state_dict(model_dict)
+    return model
+
+
+def resnet101(pretrained, num_classes, channels):
+
+    model = ResNet(Bottleneck, [3, 4, 23, 3], num_classes=num_classes, channels=channels)
+    if pretrained:
+        pretrain_dict = model_zoo.load_url(model_urls["resnet101"])
+        model_dict = model.state_dict()
+        model_dict = weight_transform(model_dict, pretrain_dict, channels)
+        model.load_state_dict(model_dict)
+
+    return model
+
+
+def resnet152(pretrained, num_classes, channels):
+
+    model = ResNet(Bottleneck, [3, 8, 36, 3], num_classes, channels)
+    if pretrained:
+        pretrain_dict = model_zoo.load_url(model_urls["resnet152"])
+        model_dict = model.state_dict()
+        model_dict = weight_transform(model_dict, pretrain_dict, channels)
+        model.load_state_dict(model_dict)
+    return model
